+++ selftest-core/src/test/java/net/oneandone/httpselftest/test/util/RequestHelperTest.java
@@ -19,11 +19,12 @@ public class RequestHelperTest {
     @Test
     public void addFormParams() throws Exception {
         StringJoiner joiner = new StringJoiner("&");
-        TestConfigs config = new TestConfigs("ka", "kö", "k&", "k=", "kunused");
-        config.put("id", "va", "vö", "v&", "v=", "vunused");
+        TestConfigs.Builder configBuilder = new TestConfigs.Builder("ka", "kö", "k&", "k=", "kunused");
+        configBuilder.put("id", "va", "vö", "v&", "v=", "vunused");
+        TestConfigs config = new TestConfigs(configBuilder);
 
         // exercise
-        RequestHelper.addFormParamsUtf8(joiner, config.getValues("id"), "ka", "k&", "kö", "k=");
+        RequestHelper.addFormParamsUtf8(joiner, config.create("id"), "ka", "k&", "kö", "k=");
 
         assertThat(joiner.toString()).isEqualTo("ka=va&k%26=v%26&k%C3%B6=v%C3%B6&k%3D=v%3D");
     }